Rename examples package and utilize armory as a namespace package
