+++ examples/pyproject.toml
@@ -58,14 +58,14 @@ allow-direct-references = true
 
 [tool.hatch.version]
 source = "vcs"
-path = "src/charmory_examples/version.py"
+path = "src/armory/examples/version.py"
 
 [tool.hatch.version.raw-options]
 root = "../"
 version_scheme = "post-release"
 
 [tool.hatch.build.hooks.vcs]
-version-file = "src/charmory_examples/version.py"
+version-file = "src/armory/examples/version.py"
 template = """
 # coding: utf-8
 # File automatically generated by setuptools_scm.
