+++ examples/src/armory/examples/image_classification/eurosat_precomputed_pgd/create_finetuned_model.py
