+++ examples/src/armory/examples/image_classification/eurosat_precomputed_pgd/run_against_pregenerated.py
@@ -1,16 +1,15 @@
 from pprint import pprint
 
-from charmory_examples.image_classification.eurosat_precomputed_pgd.evaluation import (
+from armory.examples.image_classification.eurosat_precomputed_pgd.evaluation import (
     create_evaluation_task,
     get_cli_args,
 )
-
 from charmory.engine import EvaluationEngine
 from charmory.track import track_param
 
 if __name__ == "__main__":
-    args = get_cli_args(with_attack=True)
-    task = create_evaluation_task(with_attack=True, **vars(args))
+    args = get_cli_args(with_attack=False)
+    task = create_evaluation_task(with_attack=False, **vars(args))
     track_param("main.num_batches", args.num_batches)
     engine = EvaluationEngine(task, limit_test_batches=args.num_batches)
     results = engine.run()
