+++ examples/src/armory/examples/image_classification/eurosat_precomputed_pgd/run_all.sh
