+++ examples/src/armory/examples/image_classification/food101/food101_train.py
