+++ examples/src/armory/examples/image_classification/kaludi_food_category_pgd.py
@@ -2,12 +2,12 @@ from pprint import pprint
 
 import art.attacks.evasion
 from art.estimators.classification import PyTorchClassifier
-from charmory_examples.utils.args import create_parser
 import jatic_toolbox
 import numpy as np
 import torch.nn
 from transformers.image_utils import infer_channel_dimension_format
 
+from armory.examples.utils.args import create_parser
 from armory.metrics.compute import BasicProfiler
 from charmory.data import ArmoryDataLoader
 from charmory.engine import EvaluationEngine
