+++ examples/src/armory/examples/image_classification/pokemon_pgd.py
