+++ examples/src/armory/examples/object_detection/coco_precomputed_patch/.gitignore
