+++ examples/src/armory/examples/object_detection/coco_precomputed_patch/evaluation.py
@@ -1,16 +1,19 @@
-from pprint import pprint
+"""Definition for the COCO object detection evaluation"""
+
+from typing import Optional
 
 import art.attacks.evasion
 from art.estimators.object_detection import PyTorchObjectDetector
-from charmory_examples.utils.args import create_parser
+import datasets
 import jatic_toolbox
+from jatic_toolbox.interop.huggingface import HuggingFaceObjectDetectionDataset
 import numpy as np
 from transformers import AutoImageProcessor, AutoModelForObjectDetection
 
 from armory.art_experimental.attacks.patch import AttackWrapper
+from armory.examples.utils.args import create_parser
 from armory.metrics.compute import BasicProfiler
 from charmory.data import ArmoryDataLoader
-from charmory.engine import EvaluationEngine
 from charmory.evaluation import Attack, Dataset, Evaluation, Metric, Model
 from charmory.experimental.transforms import (
     BboxFormat,
@@ -21,21 +24,22 @@ from charmory.tasks.object_detection import ObjectDetectionTask
 from charmory.track import track_init_params, track_params
 
 
-def get_cli_args():
+def get_cli_args(with_attack: bool):
     parser = create_parser(
-        description="Run COCO object detection example using models and datasets from the JATIC toolbox",
+        description="Run COCO object detection evaluation",
         batch_size=4,
         export_every_n_batches=5,
         num_batches=20,
     )
+    if not with_attack:
+        parser.add_argument(
+            "dataset_path",
+            type=str,
+        )
     return parser.parse_args()
 
 
-@track_params(prefix="main")
-def main(batch_size, export_every_n_batches, num_batches):
-    ###
-    # Model
-    ###
+def _load_model():
     model = track_params(AutoModelForObjectDetection.from_pretrained)(
         pretrained_model_name_or_path="hustvl/yolos-tiny"
     )
@@ -56,16 +60,28 @@ def main(batch_size, export_every_n_batches, num_batches):
         ),
     )
 
-    ###
-    # Dataset
-    ###
-    dataset = track_params(jatic_toolbox.load_dataset)(
-        provider="huggingface",
-        dataset_name="rafaelpadilla/coco2017",
-        task="object-detection",
-        split="val",
-        category_key="label",
+    eval_model = Model(
+        name="yolos",
+        model=detector,
     )
+
+    return eval_model, detector
+
+
+def _load_dataset(batch_size: int, dataset_path: Optional[str] = None):
+    if dataset_path is not None:
+        ds = track_params(datasets.load_from_disk)(dataset_path=dataset_path)
+        if isinstance(ds, datasets.DatasetDict):
+            ds = ds["test"]
+        dataset = HuggingFaceObjectDetectionDataset(ds, category_key="label")
+    else:
+        dataset = track_params(jatic_toolbox.load_dataset)(
+            provider="huggingface",
+            dataset_name="rafaelpadilla/coco2017",
+            task="object-detection",
+            split="val",
+            category_key="label",
+        )
 
     # Have to filter out non-RGB images
     def filter(sample):
@@ -78,9 +94,7 @@ def main(batch_size, export_every_n_batches, num_batches):
 
     dataset.set_transform(
         create_object_detection_transform(
-            # Resize and pad images to 512x512
             max_size=512,
-            # Scale to [0,1]
             float_max_value=255,
             format=BboxFormat.COCO,
             label_fields=["label", "id", "iscrowd"],
@@ -89,21 +103,19 @@ def main(batch_size, export_every_n_batches, num_batches):
 
     dataloader = ArmoryDataLoader(dataset, batch_size=batch_size)
 
-    ###
-    # Evaluation
-    ###
-    eval_dataset = Dataset(
+    return Dataset(
         name="coco",
         test_dataloader=dataloader,
         x_key="image",
         y_key="objects",
     )
 
-    eval_model = Model(
-        name="faster-rcnn-resnet50",
-        model=detector,
-    )
 
+def _create_metric():
+    return Metric(profiler=BasicProfiler())
+
+
+def _create_attack(detector: PyTorchObjectDetector):
     patch = track_init_params(art.attacks.evasion.RobustDPatch)(
         detector,
         patch_shape=(3, 50, 50),
@@ -116,40 +128,49 @@ def main(batch_size, export_every_n_batches, num_batches):
         verbose=False,
     )
 
-    eval_attack = Attack(
+    return Attack(
         name="RobustDPatch",
         attack=AttackWrapper(patch),
         use_label_for_untargeted=False,
     )
 
-    eval_metric = Metric(
-        profiler=BasicProfiler(),
-    )
+
+def create_evaluation(
+    batch_size: int,
+    with_attack: bool,
+    dataset_path: Optional[str] = None,
+    **kwargs,
+) -> Evaluation:
+    model, detector = _load_model()
+    dataset = _load_dataset(batch_size=batch_size, dataset_path=dataset_path)
+    attack = _create_attack(detector) if with_attack else None
+
+    attack_type = "generated" if with_attack else "precomputed"
 
     evaluation = Evaluation(
-        name="coco-yolos-object-detection",
-        description="COCO object detection using YOLO from HuggingFace",
-        author="",
-        dataset=eval_dataset,
-        model=eval_model,
-        attack=eval_attack,
-        metric=eval_metric,
+        name=f"coco-yolos-{attack_type}-robustdpatch",
+        description=f"COCO object detection with {attack_type} RobustDPatch attack",
+        author="TwoSix",
+        dataset=dataset,
+        model=model,
+        attack=attack,
+        metric=_create_metric(),
     )
 
-    ###
-    # Engine
-    ###
+    return evaluation
+
+
+@track_params
+def create_evaluation_task(
+    export_every_n_batches: int, with_attack: bool, **kwargs
+) -> ObjectDetectionTask:
+    evaluation = create_evaluation(with_attack=with_attack, **kwargs)
 
     task = ObjectDetectionTask(
         evaluation,
         export_every_n_batches=export_every_n_batches,
         class_metrics=False,
+        skip_attack=not with_attack,
     )
-    engine = EvaluationEngine(task, limit_test_batches=num_batches)
-    results = engine.run()
 
-    pprint(results)
-
-
-if __name__ == "__main__":
-    main(**vars(get_cli_args()))
+    return task
