+++ examples/src/armory/examples/object_detection/coco_yolos_robustdpatch.py
@@ -1,19 +1,16 @@
-"""Definition for the COCO object detection evaluation"""
-
-from typing import Optional
+from pprint import pprint
 
 import art.attacks.evasion
 from art.estimators.object_detection import PyTorchObjectDetector
-from charmory_examples.utils.args import create_parser
-import datasets
 import jatic_toolbox
-from jatic_toolbox.interop.huggingface import HuggingFaceObjectDetectionDataset
 import numpy as np
 from transformers import AutoImageProcessor, AutoModelForObjectDetection
 
 from armory.art_experimental.attacks.patch import AttackWrapper
+from armory.examples.utils.args import create_parser
 from armory.metrics.compute import BasicProfiler
 from charmory.data import ArmoryDataLoader
+from charmory.engine import EvaluationEngine
 from charmory.evaluation import Attack, Dataset, Evaluation, Metric, Model
 from charmory.experimental.transforms import (
     BboxFormat,
@@ -24,22 +21,21 @@ from charmory.tasks.object_detection import ObjectDetectionTask
 from charmory.track import track_init_params, track_params
 
 
-def get_cli_args(with_attack: bool):
+def get_cli_args():
     parser = create_parser(
-        description="Run COCO object detection evaluation",
+        description="Run COCO object detection example using models and datasets from the JATIC toolbox",
         batch_size=4,
         export_every_n_batches=5,
         num_batches=20,
     )
-    if not with_attack:
-        parser.add_argument(
-            "dataset_path",
-            type=str,
-        )
     return parser.parse_args()
 
 
-def _load_model():
+@track_params(prefix="main")
+def main(batch_size, export_every_n_batches, num_batches):
+    ###
+    # Model
+    ###
     model = track_params(AutoModelForObjectDetection.from_pretrained)(
         pretrained_model_name_or_path="hustvl/yolos-tiny"
     )
@@ -60,28 +56,16 @@ def _load_model():
         ),
     )
 
-    eval_model = Model(
-        name="yolos",
-        model=detector,
+    ###
+    # Dataset
+    ###
+    dataset = track_params(jatic_toolbox.load_dataset)(
+        provider="huggingface",
+        dataset_name="rafaelpadilla/coco2017",
+        task="object-detection",
+        split="val",
+        category_key="label",
     )
-
-    return eval_model, detector
-
-
-def _load_dataset(batch_size: int, dataset_path: Optional[str] = None):
-    if dataset_path is not None:
-        ds = track_params(datasets.load_from_disk)(dataset_path=dataset_path)
-        if isinstance(ds, datasets.DatasetDict):
-            ds = ds["test"]
-        dataset = HuggingFaceObjectDetectionDataset(ds, category_key="label")
-    else:
-        dataset = track_params(jatic_toolbox.load_dataset)(
-            provider="huggingface",
-            dataset_name="rafaelpadilla/coco2017",
-            task="object-detection",
-            split="val",
-            category_key="label",
-        )
 
     # Have to filter out non-RGB images
     def filter(sample):
@@ -94,7 +78,9 @@ def _load_dataset(batch_size: int, dataset_path: Optional[str] = None):
 
     dataset.set_transform(
         create_object_detection_transform(
+            # Resize and pad images to 512x512
             max_size=512,
+            # Scale to [0,1]
             float_max_value=255,
             format=BboxFormat.COCO,
             label_fields=["label", "id", "iscrowd"],
@@ -103,19 +89,21 @@ def _load_dataset(batch_size: int, dataset_path: Optional[str] = None):
 
     dataloader = ArmoryDataLoader(dataset, batch_size=batch_size)
 
-    return Dataset(
+    ###
+    # Evaluation
+    ###
+    eval_dataset = Dataset(
         name="coco",
         test_dataloader=dataloader,
         x_key="image",
         y_key="objects",
     )
 
+    eval_model = Model(
+        name="faster-rcnn-resnet50",
+        model=detector,
+    )
 
-def _create_metric():
-    return Metric(profiler=BasicProfiler())
-
-
-def _create_attack(detector: PyTorchObjectDetector):
     patch = track_init_params(art.attacks.evasion.RobustDPatch)(
         detector,
         patch_shape=(3, 50, 50),
@@ -128,49 +116,40 @@ def _create_attack(detector: PyTorchObjectDetector):
         verbose=False,
     )
 
-    return Attack(
+    eval_attack = Attack(
         name="RobustDPatch",
         attack=AttackWrapper(patch),
         use_label_for_untargeted=False,
     )
 
-
-def create_evaluation(
-    batch_size: int,
-    with_attack: bool,
-    dataset_path: Optional[str] = None,
-    **kwargs,
-) -> Evaluation:
-    model, detector = _load_model()
-    dataset = _load_dataset(batch_size=batch_size, dataset_path=dataset_path)
-    attack = _create_attack(detector) if with_attack else None
-
-    attack_type = "generated" if with_attack else "precomputed"
-
-    evaluation = Evaluation(
-        name=f"coco-yolos-{attack_type}-robustdpatch",
-        description=f"COCO object detection with {attack_type} RobustDPatch attack",
-        author="TwoSix",
-        dataset=dataset,
-        model=model,
-        attack=attack,
-        metric=_create_metric(),
+    eval_metric = Metric(
+        profiler=BasicProfiler(),
     )
 
-    return evaluation
+    evaluation = Evaluation(
+        name="coco-yolos-object-detection",
+        description="COCO object detection using YOLO from HuggingFace",
+        author="",
+        dataset=eval_dataset,
+        model=eval_model,
+        attack=eval_attack,
+        metric=eval_metric,
+    )
 
-
-@track_params
-def create_evaluation_task(
-    export_every_n_batches: int, with_attack: bool, **kwargs
-) -> ObjectDetectionTask:
-    evaluation = create_evaluation(with_attack=with_attack, **kwargs)
+    ###
+    # Engine
+    ###
 
     task = ObjectDetectionTask(
         evaluation,
         export_every_n_batches=export_every_n_batches,
         class_metrics=False,
-        skip_attack=not with_attack,
     )
+    engine = EvaluationEngine(task, limit_test_batches=num_batches)
+    results = engine.run()
 
-    return task
+    pprint(results)
+
+
+if __name__ == "__main__":
+    main(**vars(get_cli_args()))
