+++ examples/src/armory/examples/object_detection/xview/xview_fasterrcnn_robustdpatch.py
