--- library/src/armory/__init__.py
@@ -1,4 +1,0 @@
-# Adversarial Robustness Evaluation Test Bed (ARMORY)
-
-# TODO: drop this import which was justt for backwards compatibility
-from armory.logs import log
